Feat android response models

+++ src/SDK/Language/Kotlin.php
@@ -97,7 +97,7 @@ class Kotlin extends Language {
     {
         switch ($type) {
             case self::TYPE_INTEGER:
-                return 'Int';
+                return 'Long';
             case self::TYPE_NUMBER:
                 return 'Double';
             case self::TYPE_STRING:
